Move files around for refactoring
