+++ lib/formats/index.js
@@ -1,0 +1,5 @@
+var html = require("./html");
+
+module.exports = {
+  html: html
+}
